+++ src/react/components/autocompleter/__tests__/xui-autocompleter-test.js
@@ -34,9 +34,9 @@ describe('XUIAutocompleter', () => {
 	});
 
 	it('has data-automationid set on the input, list and container', () => {
-		expect(wrapper.find({'data-automationid':'baseAC-input'})).toHaveLength(1);
-		expect(wrapper.find({'data-automationid':'baseAC-input--container'})).toHaveLength(1);
-		expect(wrapper.instance().ddt.props.dropdown.props.qaHook).toEqual('baseAC-list');
+		expect(wrapper.find({'data-automationid':'baseAC--input'})).toHaveLength(1);
+		expect(wrapper.find({'data-automationid':'baseAC--container'})).toHaveLength(1);
+		expect(wrapper.instance().ddt.props.dropdown.props.qaHook).toEqual('baseAC--list');
 	});
 
 	it('inserts the searchValue inside the input', () => {