+++ src/react/components/banner/XUIBannerAction.js
@@ -5,12 +5,11 @@ import cn from 'classnames';
 
 export default function XUIBannerAction(props) {
 	const className = cn(props.className, 'xui-button-small');
-	const isLink = !!props.href;
-	const buttonQaHook = props.qaHook && `${props.qaHook}-button`;
+	const buttonQaHook = props.qaHook && `${props.qaHook}--button`;
 
 	return (
 		<li className="xui-banner--action" data-automationId={props.qaHook}>
-			<XUIButton {...props} isLink={isLink} variant="link" className={className} qaHook={buttonQaHook}>{props.children}</XUIButton>
+			<XUIButton href={props.href} isLink={props.isLink} variant="link" className={className} qaHook={buttonQaHook}>{props.children}</XUIButton>
 		</li>
 	);
 }
@@ -19,5 +18,12 @@ XUIBannerAction.propTypes = {
 	className: PropTypes.string,
 	children: PropTypes.node,
 	qaHook: PropTypes.string,
-	href: PropTypes.string
+	/** Target URl of action, will also trun the button into a link, <a /> tag. */
+	href: PropTypes.string,
+	/** Whether or not to render this button using an tag. */
+	isLink: PropTypes.bool
 };
+
+XUIBannerAction.defaultProps = {
+	isLink: false
+}